test(do): prove fresh and migrated objects reach the same schema
Durable Objects have no migration runner, so an object created today and one
that predates devices arrive by different routes. The divergence that matters
is silent — a column NOT NULL on one and nullable on the other — so the two
are now compared directly rather than assumed equal.

Also asserts what the '' sentinel exists for: two writes to the default
device produce one row. With NULL they would produce two, because SQLite
treats NULLs as distinct in a unique key.

The schema ladder moves to its own module so a test can drive it without
pulling in the Durable Object runtime.
